backend/films_backend/requirements.txt
Django==5.1.2
gunicorn
dj-database-url
psycopg2-binary
whitenoise
django-cors-headers==4.4.0
django-taggit==6.1.0
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.1
asgiref==3.8.1